migrate note
